requirements.txt
rumps==0.4.0
pyaudio==0.2.13
sounddevice==0.4.6
numpy==1.24.3
llama-cpp-python>=0.3.0
python-dotenv==1.0.0
pyannote.audio>=3.1.0
torch>=2.0.0
pytest>=7.0.0
